Fixup compiler warnings

+++ plugins/wdc/wdc-nvme.c
@@ -5406,7 +5406,7 @@ static int wdc_drive_status(int argc, char **argv, struct command *command,
 		fprintf(stderr, "ERROR : WDC : Get Format Corrupt Reason Failed\n");
 
 	printf("  Drive Status :- \n");
-	if (le32_to_cpu(eol_status) >= 0) {
+	if ((int)le32_to_cpu(eol_status) >= 0) {
 		printf("  Percent Used:				%"PRIu32"%%\n",
 				le32_to_cpu(eol_status));
 	}
